backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "qave"
version = "0.1.0"
description = "Quantum algorithm visualization engine for trace and animation generation"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "QAVE Contributors" }]
keywords = ["quantum", "visualization", "qiskit", "education", "animation"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
  "Pillow>=4.2.1",
  "numpy>=2.0.0,<3.0.0",
  "qiskit[visualization]>=1.1.0,<2.0.0",
  "pydantic>=2.7.0,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/q-inho/qave"
Repository = "https://github.com/q-inho/qave"

[project.optional-dependencies]
dev = [
  "coverage>=7.6.0,<8.0.0",
  "mypy>=1.11.0,<2.0.0",
  "pre-commit>=4.0.0,<5.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.6.0,<1.0.0",
  "typos>=1.35.0,<2.0.0",
]

[project.scripts]
qave-export-schemas = "qave_backend.contracts.schema_export:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qave", "src/qave_backend"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q --capture=no"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
show_error_codes = true

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"
